+++ database/repositories/tyre_impression_repository.py
@@ -1,0 +1,10 @@
+from database.models import TyreImpression
+from database.repositories.base_repository import BaseRepository
+
+
+class TyreImpressionRepository(BaseRepository[TyreImpression]):
+
+    def __init__(self):
+        super().__init__(TyreImpression)
+
+