+++ tests/mocks/services/__init__.py
@@ -1,1 +1,3 @@
 from .mock_file_service import MockFileService
+from .mock_tyre_model_service import MockTyreModelService
+from .mock_tyre_impression_service import MockTyreImpressionService