+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Added
 - Created a tyre impression model with a status enum data type
+- Created database migration for new `tyre_impressions` table
 
 ## v0.0.4 21/05/2026
 ### Test Coverage